rf: add implicit pragma to ex
The "implicit" pragma allows matching expressions based on the type of
the implicit variable they're being assigned to. For example, if toint
has been declared as "var toint func(int)", then "toint(x + 1)" will
only match in contexts where "x + 1" is being assigned to a variable
of type int, passed to an argument of type, etc.