rf: generalize substitution validation
Prevents substitutions of expressions on the RHS side of an N:1
assignment when the substition expression isn't appropriate. In
particular, this allows rewrites that substitute map indexing with
separate functions to handle 1-value and 2-value forms:

    # rf will automatically pick the appropriate substition
    # based on context.
    m[k] -> m.Get(k)
    m[k] -> m.Get2(k)

Caveat: Make sure to list the single-value substition first, because
contextAppropriate does not yet prevent multi-value substitions in
single-value contexts. So for now, scripts must take advantage of
short-circuiting semantics instead.