rf: add export command
This is really just a shorthand for `mv sym exportedSym` but it
makes it more convenient when exporting a slew
of symbols such as a bunch of constants or types.

In the future, it would be nice to have support for renaming all
symbols in a particular range (for example to export all constants
in a particular const block), but this is hopefully a useful start.