rf: robust detection of the current module
rf only modifies files in the current module, but it has to load
imported modules for type checking, so it records on every package
whether it's in the current module or not. However, currently, it uses
the DepOnly field from "go list" to determine this. It's true that if
.DepOnly is false, then it's in the current module, but there are
various conditions under which .DepOnly is true and the package is
still in the current module (these conditions may not always be right,
but they're complicated regardless; see golang/go#60794).

Inside of using .DepOnly, use .Module.Version, which is nil for
anything in the current workspace, and thus much more closely matches
the semantics of "can we edit this?"