rf: use multiple Snapshots to resolve testing cycles
rf substitutes "p [p.test]" packages for "p" packages because the "p
[p.test]" package is a strict superset of the source files of "p",
allowing it to cover more source code. (This substitution is not
strictly sound, but works for almost all code in practice.) However,
"p [p.test]" can also contain more imports than "p", so when it does
this on all packages in a module at once, it's possible to introduce
cycles in the import graph.

For example, if you have two packages "p" and "q" that have no import
relationship, but p/p_test.go imports "q" and q/q_test.go imports "p",
then expanding both to including their test files will result in a
cycle.

Currently, rf uses some heuristics to try to cull this, but generally
just fails to type-check the module.

This commit uses the new ability to refactor multiple Snapshots in
parallel to correctly address this issue. It introduces a heuristic
algorithm for packages to include their tests that creates multiple
Snapshots when necessary to avoid cycles. Despite the simplicity of
the heuristic, it results in only two Snapshots for std.

This change has other benefits as well. It introduces a package graph
abstraction separate from Snapshot that lets us abstract away various
operations on the package graph.

Also, since Snapshot creation now flattens away package IDs, we can
use package paths almost exclusively through the rest of the code (the
one remaining exception is the need to consult the ImportMap for
vendored packages in std, but that's simpler than before).