rf: maintain the current Snapshot in Refactor
Currently it's up to the user of Refactor to track the current
Snapshot. We're about to switch to tracking multiple Snapshots that
must converge, so this is going to get trickier. Simplify things by
tracking the current Snapshot in Refactor, where we can more easily
switch to tracking more than one. This also switches around the API
for accessing the current Snapshot: the new Refactor.Snapshot returns
the current Snapshot (performing the initial load if necessary) and
the new Refactor.Commit finalizes the current Snapshot to create a new
Snapshot.