rf: fix several bugs in CreateFile
There seemed to be broad inconsistency on how Snapshot dealt with
newly created files. The root inconsistency was that most things
treated Snapshot.files as a snapshot of the initial state of all
files, where any modifications were kept in edits. However, this was
*not* true when a new file was created by CreateFile, which added the
new file to both edits and files.

This change starts by only adding the new file to edits.

However, this revealed several other issues and complexities that
arose from this inconsistency. CreateFile also did not create a file
name in the "short path" form that was expected by the rest of the
code, so the key used in Snapshot.edits was wrong. We fix that by
simplifying. CreateFile now calls buildCache.newTextFile to construct
the File rather than making the File on its own, and we move the logic
to construct a proper short path from newFile to newTextFile so it's
shared.

This incorrect key in edits revealed an issue in Refactor.Apply (which
was Snapshot.Load until recently), where it attempted to gather all
new files by directory, but never successfully consumed this map
because the keys were wrong. This turned out to be okay because it was
already getting the files from Package.Files, so if it had consumed
this map, it would have added newly created files twice. Hence, we
simply delete the logic to create this map because it's not
necessarily.

Finally, Snapshot.Diff did not look for newly created files the way
that Snapshot.Write did, so it would miss these if they hadn't already
been Snapshot.Apply'd. So we fix Diff to look for new files like Write
does.