rf: move error handling into new ErrorList type
Currently error handling is bundled into Snapshot and errors are
emitted to stderr eagerly. We're going to have to do more error
handling outside the scope of any specific Snapshot, and we're going
to need to handle multiple error lists for parallel Snapshots. So,
this CL introduces an ErrorList type as a first-class list of errors.
This is returned and wrapped just like a regular Go error. Writing
these errors to stderr is now the responsibility of the application
layer (though we still flush it to stderr if we panic in key places
inside the refactor package, since otherwise key information about
that panic may be lost).