rf: support commands that only work on some Snapshots
With different Snapshots where some include test packages for the
target and some don't, it's possible for commands to only work on
Snapshot including the test package (for example, if it references
something in the test package). Currently refactoring will fail unless
the command works on all Snapshots. With this commit, we permit a
command to fail with an errPrecondition, as long as it succeeds on *at
least one* Snapshot.