rf: support refactoring across GOOS/GOARCHes
This creates Snapshots for every GOOS/GOARCH so that refactoring can
be applied across all of them.

This is, unfortunately, quite slow and requires a large amount of RAM.
The actual refactoring step doesn't take long, but the initial "go
list -export" can take an extremely long time as it rebuilds std for
every GOOS/GOARCH, and then rf has to re-type-check all of the
per-GOOS/GOARCH package graphs after every refactoring step, even
though the vast majority of that is duplicate work. Because of all of
this, we make this functionality opt-in. (I believe that if we used
package IDs instead of parallel Snapshot and had fine-grained type
check cache invalidation, this could be dramatically more efficient.)