rf: only load necessary export data
Currently, rf asks for the export data of all packages, but then
ignores it for most (possibly all?) packages. Creating the export data
requires compiling everything, so this is a huge amount of wasted
work.

Fix this by doing two passes, where the first loads the whole package
graph and figures out what export data we need, and the second gets
just the export data we need.