git-generate: clean gitdir filepath on Windows (#32)
For some reason, on Windows, git prints the path with '/' separators
instead of '\'. This by itself isn't a problem, but it intersects
poorly with prefix-based relative path computation in walkGit:

	cannot compute relative path: C:/reporoot vs C:\reporoot\file.txt

We could solve that by using filepath.Rel, but since cleaning the path
is all it takes to let the prefix-based computation work on Windows too,
pick that for now.

Tested on a Windows machine.

Fixes #30.