git-generate: modernize slightly
The go1.16 build constraint was added back when git-generate was tested
with Go 1.15 and 1.16, and its intention was to raise the minimum to as
high as 1.16.

By now, the go directive in go.mod is already at 1.20 and will only get
higher, so the go1.16 build constraint is no longer needed, and keeping
it around will begin to work counter to its original intention. Drop it.

Also apply gofmt and migrate away from the io/ioutil package while here.