rf: add test for removing via text ranges
This change adds a test for the `rm` command, wherein the removal is
done by specifying a text range.

Documentation is also updated to reflect that the `rm` command
implements the removal of struct fields (see `testdata/rm_field.txt`),
and the removal of interface methods (see `testdata/rm_iface.txt`).