rf: improve detection of `unsafe` special case
When using profile guided optimization, the PkgPath can contain a
suffix with the binary in question.  For example, the PkgPath:
  `unsafe [cmd/compile]`
was encountered when executing rf in the cmd/go subtree.  This commit
uses a regex to better identify when the special case of the `unsafe`
package should apply.  Note that we need to be cautious not to falsely
match a package such as `unsafeheader`.